No more flaffing about; Exceptions are always fatal
I've been going back and forth on this a lot and have decided to commit to exceptions always being fatal for the time being. This may yet change in the future but for now I'm not going to worry about exceptions.

This change is partly induced by the realization that there is absolutely no reason to make a breakpoint exception - the VM itself is used as a debugger and there is no need to hook in with some code or process external to the program being executed.

This commit causes the "ebreak" instruction to be recognized by the debugger (stop there after entering "resume") and to no longer create an exception. The "rfe" (return from exception) instruction has been removed.

Also, incidentally, the exception code previously reserved for breakpoints is now used to indicate that the PC fell in memory not considered to be executable, and now only the text segment is executable. Maybe one day later running code outside the text segment may be a thing, but today is not that day.