Fixes to & tests for some pseudo-instructions
Compiler is now smart enough to emit an error if the arguments given to a pseudo-instruction will result in (presumably) unexpected behavior, e.g. using the same rd and rs1 for addwi (doesn't work because rd has to be modified before accessing rs1 - so if they're the same, the value in rs1 is lost before getting around to reading it. Might be nice to come up with a generalization where the pseudo-instruction can somehow still work with just the one register, but that seems implausible)