Revise lb, lbu, lh, lhu, lw: Addr is rs2+rd instead of rs1+rd
This change was made such that the same register is added to the immediate value to produce an address for both stores and loads.

This option was chosen in favor of reversing sb, sh, sw register arguments because I like that the address register is adjacent to the offset immediate value in the argument list, and it is easier and simpler given how the assembler is written to change the load instructions' register than it would be to swap the store instructions' registers while maintaining that nice syntax.

I guess it's slightly funky that now there are a few instructions that use rs2 but not rs1, but it's a very small thing and I feel like keeping the assembler's argument parsing logic simpler outweighs that very small amount of funkiness.