More shuffling around of rs1, rs2 ordering
- Undo earlier changes to load instructions - address is rs1 + imm again
- Change store instructions instead. The syntax is the same, but now the parser expects rs2 first and then rs1, and the meaning in the interpreter is switched around

Now, consistently, every instruction that adds the immediate to a source register adds it to the first source register. In theory this should simplify hardware implementations, should they ever exist.

One motivation for choosing rs1 over rs2 for this situation, as I'd initially switched before (forgetting jalr notwithstanding) is that it opens the possibility of using the 3 bits of rs2 as an additional 3 high bits of the immediate value, for instructions that would not otherwise be using rs2. Ditto instructions that have an immediate but recognize neither rs1 nor rs2. I don't know if there will ever be a practical reason to do this, since even if it allows more bits to fit into some instructions' immediate values it would add complexity and inconsistency to the ISA, but it seems prudent to leave the option open - and using those extra register bits for something will surely be easier if they are adjacent to the immediate value and not separated from the immedate by bits representing a register that is used.